Refactorings; holiday and lecture categories
There's now an EventType for the Lecture's (either Lecture, Exam or
Holiday).

Implemented parsing for edge case with location data in the event name.

Data is now serialized to a separate struct for serde at first and then
the various checks and fixes are applied to the data and stored for
later use.